# backend/app/api/static/docs/swagger.yaml
basePath: /api
definitions:
...
  repo.EntityListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/repo.EntitySummary'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPrice:
        type: number
    type: object
...
  repo.ItemAttachmentUpdate:
    properties:
      primary:
        type: boolean
      title:
        type: string
      type:
        type: string
    type: object
...
  repo.PaginationResult-repo_EntitySummary:
    properties:
      items:
        items:
          $ref: '#/definitions/repo.EntitySummary'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  v1.ActionAmountResult:
    properties:
      completed:
        type: integer
    type: object
...
  v1.WipeInventoryOptions:
    properties:
      wipeLocations:
        type: boolean
      wipeMaintenance:
        type: boolean
      wipeTags:
        type: boolean
    type: object
...
  validate.ErrorResponse:
    properties:
      error:
        type: string
      fields:
        type: string
    type: object
info:
  contact:
    name: Homebox Team
    url: https://discord.homebox.software
  description: Track, Manage, and Organize your Things.
  title: Homebox API
  version: "1.0"
paths:
  /v1/actions/create-missing-thumbnails:
    post:
      description: Creates thumbnails for items that are missing them
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Create Missing Thumbnails
      tags:
      - Actions
  /v1/actions/ensure-asset-ids:
    post:
      description: Ensures all items in the database have an asset ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Ensure Asset IDs
      tags:
      - Actions
  /v1/actions/ensure-import-refs:
    post:
      description: Ensures all items in the database have an import ref
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Ensures Import Refs
      tags:
      - Actions
  /v1/actions/set-primary-photos:
    post:
      description: Sets the first photo of each item as the primary photo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Set Primary Photos
      tags:
      - Actions
  /v1/actions/wipe-inventory:
    post:
      description: Deletes all items in the inventory
      parameters:
      - description: Wipe options
        in: body
        name: options
        schema:
          $ref: '#/definitions/v1.WipeInventoryOptions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Wipe Inventory
      tags:
      - Actions
  /v1/actions/zero-item-time-fields:
    post:
      description: Resets all item date fields to the beginning of the day
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Zero Out Time Fields
      tags:
      - Actions
  /v1/assets/{id}:
    get:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.PaginationResult-repo_EntitySummary'
      security:
      - Bearer: []
      summary: Get Item by Asset ID
      tags:
      - Items
...
  /v1/entities:
    get:
      parameters:
      - description: search string
        in: query
        name: q
        type: string
      - description: page number
        in: query
        name: page
        type: integer
      - description: items per page
        in: query
        name: pageSize
        type: integer
      - collectionFormat: multi
        description: tags Ids
        in: query
        items:
          type: string
        name: tags
        type: array
      - collectionFormat: multi
        description: parent Ids
        in: query
        items:
          type: string
        name: parentIds
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.EntityListResult'
      security:
      - Bearer: []
      summary: Query All Entities
      tags:
      - Entities
    post:
      parameters:
      - description: Entity Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.EntityCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repo.EntityOut'
      security:
      - Bearer: []
      summary: Create Entity
      tags:
      - Entities
  /v1/entities/{id}:
    delete:
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Delete Entity
      tags:
      - Entities
    get:
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.EntityOut'
      security:
      - Bearer: []
      summary: Get Entity
      tags:
      - Entities
    patch:
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Entity Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.EntityPatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.EntityOut'
      security:
      - Bearer: []
      summary: Patch Entity
      tags:
      - Entities
    put:
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Entity Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.EntityUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.EntityOut'
      security:
      - Bearer: []
      summary: Update Entity
      tags:
      - Entities
  /v1/entities/{id}/attachments:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: File attachment
        in: formData
        name: file
        required: true
        type: file
      - description: Type of file
        in: formData
        name: type
        type: string
      - description: Is this the primary attachment
        in: formData
        name: primary
        type: boolean
      - description: name of the file including extension
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repo.EntityOut'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/validate.ErrorResponse'
      security:
      - Bearer: []
      summary: Create Entity Attachment
      tags:
      - Entities Attachments
  /v1/entities/{id}/attachments/{attachment_id}:
    delete:
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Attachment ID
        in: path
        name: attachment_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Delete Entity Attachment
      tags:
      - Entities Attachments
    get:
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Attachment ID
        in: path
        name: attachment_id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - Bearer: []
      summary: Get Entity Attachment
      tags:
      - Entities Attachments
    put:
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Attachment ID
        in: path
        name: attachment_id
        required: true
        type: string
      - description: Attachment Update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.ItemAttachmentUpdate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.EntityOut'
      security:
      - Bearer: []
      summary: Update Entity Attachment
      tags:
      - Entities Attachments
  /v1/entities/{id}/attachments/external:
    post:
      consumes:
      - application/json
      description: Links an entity to a document or URL in an external system without
        copying
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: External document reference
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.externalAttachmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repo.EntityOut'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validate.ErrorResponse'
      security:
      - Bearer: []
      summary: Create External Link Attachment
      tags:
      - Entities Attachments
  /v1/entities/{id}/duplicate:
    post:
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Duplicate Options
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.DuplicateOptions'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repo.EntityOut'
      security:
      - Bearer: []
      summary: Duplicate Entity
      tags:
      - Entities
  /v1/entities/{id}/maintenance:
    get:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - enum:
        - scheduled
        - completed
        - both
        in: query
        name: status
        type: string
        x-enum-varnames:
        - MaintenanceFilterStatusScheduled
        - MaintenanceFilterStatusCompleted
        - MaintenanceFilterStatusBoth
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.MaintenanceEntryWithDetails'
            type: array
      security:
      - Bearer: []
      summary: Get Maintenance Log
      tags:
      - Item Maintenance
    post:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Entry Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.MaintenanceEntryCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repo.MaintenanceEntry'
      security:
      - Bearer: []
      summary: Create Maintenance Entry
      tags:
      - Item Maintenance
  /v1/entities/{id}/path:
    get:
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.EntityPath'
            type: array
      security:
      - Bearer: []
      summary: Get the full path of an entity
      tags:
      - Entities
  /v1/entities/export:
    get:
      responses:
        "200":
          description: text/csv
          schema:
            type: string
      security:
      - Bearer: []
      summary: Export Entities
      tags:
      - Entities
  /v1/entities/external/{id}:
    get:
      parameters:
      - description: External ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.EntityOut'
      security:
      - Bearer: []
      summary: Get Entity by External ID
      tags:
      - Entities
  /v1/entities/fields:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - Bearer: []
      summary: Get All Custom Field Names
      tags:
      - Entities
  /v1/entities/fields/values:
    get:
      parameters:
      - description: Field name
        in: query
        name: field
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - Bearer: []
      summary: Get All Custom Field Values
      tags:
      - Entities
  /v1/entities/import:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: CSV file to upload
        in: formData
        name: csv
        required: true
        type: file
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Import Entities
      tags:
      - Entities
  /v1/entities/tree:
    get:
      parameters:
      - description: include items in response tree
        in: query
        name: withItems
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.TreeItem'
            type: array
      security:
      - Bearer: []
      summary: Get Locations Tree
      tags:
      - Entities
  /v1/entity-types:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.EntityTypeSummary'
            type: array
      security:
      - Bearer: []
      summary: Get All Entity Types
      tags:
      - Entity Types
    post:
      parameters:
      - description: Entity Type Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.EntityTypeCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repo.EntityTypeSummary'
      security:
      - Bearer: []
      summary: Create Entity Type
      tags:
      - Entity Types
  /v1/entity-types/{id}:
    delete:
      parameters:
      - description: Entity Type ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Delete Entity Type
      tags:
      - Entity Types
    put:
      parameters:
      - description: Entity Type ID
        in: path
        name: id
        required: true
        type: string
      - description: Entity Type Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.EntityTypeUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.EntityTypeSummary'
      security:
      - Bearer: []
      summary: Update Entity Type
      tags:
      - Entity Types
  /v1/group/exports:
    get:
      description: Returns export job rows for the caller's group, newest first.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Results-repo_ExportOut'
      security:
      - Bearer: []
      summary: List Collection Exports
      tags:
      - Group
    post:
      description: Creates a pending export row and enqueues the build job. Poll the
        listing endpoint or watch the WebSocket for completion.
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/repo.ExportOut'
      security:
      - Bearer: []
      summary: Start a Collection Export
      tags:
      - Group
  /v1/group/exports/{id}:
    delete:
      description: Deletes the export row and its blob artifact.
      parameters:
      - description: Export ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Delete an Export
      tags:
      - Group
    get:
      parameters:
      - description: Export ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.ExportOut'
      security:
      - Bearer: []
      summary: Get an Export
      tags:
      - Group
  /v1/group/exports/{id}/download:
    get:
      parameters:
      - description: Export ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/zip
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - Bearer: []
      summary: Download an Export Artifact
      tags:
      - Group
  /v1/group/import:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a collection-export zip and enqueues the import job. The
        destination group must be empty. Returns the tracked import row so clients
        can poll for progress.
      parameters:
      - description: Export zip
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/repo.ExportOut'
      security:
      - Bearer: []
      summary: Import a Collection Zip
      tags:
      - Group
...
  /v1/groups/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.Group'
            type: array
      security:
      - Bearer: []
      summary: Get All Groups
      tags:
      - Group
...
  /v1/groups/members:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.UserSummary'
            type: array
      security:
      - Bearer: []
      summary: Get All Group Members
      tags:
      - Group
...
  /v1/users/logout:
    post:
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: User Logout
      tags:
      - Authentication
...
schemes:
- https
- http
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"